+++ CHANGELOG.md
@@ -10,8 +10,8 @@ and this project adheres to
 
 - @cosmjs/amino: Add `encodeEd25519Pubkey` analogue to the existing
   `encodeSecp256k1Pubkey`.
-- @cosmjs/proto-signing: Add Ed25519 support to `anyToSinglePubkey` and make
-  export `anyToSinglePubkey`.
+- @cosmjs/proto-signing: Add Ed25519 support to `encodePubkey` and
+  `anyToSinglePubkey`. Export `anyToSinglePubkey`.
 - @cosmjs/utils: Add `isDefined` which checks for `undefined` in a
   TypeScript-friendly way.